[CCUBE-1522][JY] Rename schemaHelper. Update base schema helper.
