+++ src/yup-schemas/base-schema-helper.ts
@@ -4,9 +4,10 @@ import {
     TElementMap,
 } from "src/context-providers";
 import * as yup from "yup";
-import { CONDITIONAL_RENDERING_SCHEMA, PREFILL_YUP_SCHEMA } from "./common";
 
 const ID_REGEX = /^[a-z]+(?:[A-Z0-9][a-z0-9]*)*(?:[-_][a-z0-9]+)*$/gm;
+export const PREFILL_ACTIONID_REGEX = /^[a-zA-Z0-9_-]+$/;
+export const PREFILL_PATH_REGEX = /^[a-zA-Z0-9._-]+$/;
 
 export namespace BaseSchemaHelper {
     export const getBaseSchema = (
@@ -36,8 +37,34 @@ export namespace BaseSchemaHelper {
                         .filter((id) => id !== focusedElement?.element.id),
                     "ID must not be duplicated."
                 ),
-            prefill: PREFILL_YUP_SCHEMA,
-            conditionalRendering: CONDITIONAL_RENDERING_SCHEMA,
+            prefill: yup.array().of(
+                yup.object().shape({
+                    prefillMode: yup.string().required("Source required."),
+                    actionId: yup.string().when("prefillMode", {
+                        is: "Previous source",
+                        then: (rule) =>
+                            rule
+                                .required("Action ID required.")
+                                .matches(
+                                    PREFILL_ACTIONID_REGEX,
+                                    "Invalid action ID."
+                                ),
+                        otherwise: (rule) => rule.optional(),
+                    }),
+                    path: yup
+                        .string()
+                        .required("Path required.")
+                        .matches(PREFILL_PATH_REGEX, "Invalid path."),
+                })
+            ),
+            conditionalRendering: yup.array().of(
+                yup.object().shape({
+                    fieldKey: yup.string().required("Reference required."),
+                    comparator: yup.string().required("Comparator required."),
+                    value: yup.string().required("Reference value required."),
+                    internalId: yup.string(),
+                })
+            ),
         });
     export type TBaseSchema = yup.InferType<ReturnType<typeof getBaseSchema>>;
 }
