--- src/yup-schemas/common/index.ts
@@ -1,2 +1,0 @@
-export * from "./conditional-rendering-yup-schema";
-export * from "./prefill-yup-schema";
