+++ src/yup-schemas/index.ts
@@ -1,2 +1,2 @@
-export * from "./helper";
+export * from "./schema-builder";
 export * from "./types";
