+++ src/yup-schemas/schema-builder.ts
@@ -1,12 +1,12 @@
 import { useBuilder } from "src/context-providers";
 import * as yup from "yup";
 import { EElementType } from "../context-providers/builder/element.types";
-import { BaseSchemaHelper } from "./base-helper";
+import { BaseSchemaHelper } from "./base-schema-helper";
 import { OPTION_GROUP_BASED_SCHEMA } from "./option-group-based-fields";
 import { TEXT_AREA_SCHEMA, TEXT_BASED_SCHEMA } from "./text-based-fields";
 import { TYupSchema } from "./types";
 
-export namespace SchemaHelper {
+export namespace SchemaBuilder {
     export const buildSchema = (type: EElementType): TYupSchema => {
         try {
             const { elements, focusedElement } = useBuilder();
